Ensure that ScoreAdjustments and other FeedItems are only counted once
There is a strange bug happening for some teams on the ECTF scoreboard dump, where if you perform a ScoreAdjustment on them it happens multiple times. This ensures each teams score is only added once.

The double LEFT JOIN is the source of the problem, however we need the LEFT JOIN to get all the data we need. Calling DISTINCT fixes the issue by getting rid of the duplicate entries.